Added ffmpeg-scripts for ipad vid coding
I added a set of scripts that capture the knowledge gained from a ton of
googling around and scratching my head. The fetch scripts check out the
indicated project, the configure scripts capture the configurations options
for the indicated project, and finally, the ipad_encode script runs ffmpeg on
the desired video file to format it for the ipad. This will probably need to be
changed around a bit to suit individual needs.